refactor(appointment, admin): standradize API

+++ src/main/java/com/emmkay/infertility_system_api/modules/appointment/service/AppointmentService.java
@@ -1,8 +1,7 @@
 package com.emmkay.infertility_system_api.modules.appointment.service;
 
-import com.emmkay.infertility_system_api.modules.appointment.projection.AppointmentInNext7DayProjection;
+import com.emmkay.infertility_system_api.modules.appointment.projection.AppointmentBasicProjection;
 import com.emmkay.infertility_system_api.modules.appointment.dto.request.*;
-import com.emmkay.infertility_system_api.modules.appointment.dto.response.AppointmentInNext7DayResponse;
 import com.emmkay.infertility_system_api.modules.appointment.dto.response.AppointmentResponse;
 import com.emmkay.infertility_system_api.modules.shared.exception.AppException;
 import com.emmkay.infertility_system_api.modules.shared.exception.ErrorCode;
@@ -14,6 +13,7 @@ import com.emmkay.infertility_system_api.modules.doctor.repository.DoctorReposit
 import com.emmkay.infertility_system_api.modules.reminder.repository.ReminderRepository;
 import com.emmkay.infertility_system_api.modules.schedule.entity.WorkSchedule;
 import com.emmkay.infertility_system_api.modules.schedule.repository.WorkScheduleRepository;
+import com.emmkay.infertility_system_api.modules.shared.security.CurrentUserUtils;
 import com.emmkay.infertility_system_api.modules.treatment.entity.TreatmentRecord;
 import com.emmkay.infertility_system_api.modules.treatment.entity.TreatmentStep;
 import com.emmkay.infertility_system_api.modules.treatment.repository.TreatmentRecordRepository;
@@ -24,20 +24,17 @@ import com.emmkay.infertility_system_api.modules.reminder.service.ReminderServic
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-
-@Slf4j
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -53,10 +50,7 @@ public class AppointmentService {
     ReminderRepository reminderRepository;
     TreatmentRecordRepository treatmentRecordRepository;
 
-    private Appointment isAvailableForChange(Long appointmentId, LocalDate dateChange, String shiftChange) {
-        Appointment appointment = appointmentRepository.findById(appointmentId)
-                .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
-
+    void validateAppointmentAvailableForChange(Appointment appointment, LocalDate dateChange, String shiftChange) {
         // Chỉ cho đổi trong 14 ngày tới
         LocalDate today = LocalDate.now();
         if (dateChange.isBefore(today) || dateChange.isAfter(today.plusDays(14))) {
@@ -74,30 +68,74 @@ public class AppointmentService {
                 || "CANCELLED".equalsIgnoreCase(appointment.getStatus())) {
             throw new AppException(ErrorCode.CAN_NOT_BE_UPDATED_STATUS);
         }
-        return appointment;
     }
 
-    public List<AppointmentInNext7DayResponse> getAppointInNext7Day(String doctorId) {
-        LocalDate startDate = LocalDate.now();
-        LocalDate endDate = LocalDate.now().plusDays(6);
-        List<AppointmentInNext7DayProjection> tmp = appointmentRepository.getAppointmentInNext7Day(doctorId, endDate);
+    void validateCanChangeAppointment(Appointment appointment) {
+        String scope = CurrentUserUtils.getCurrentScope();
+        String currentUserId = CurrentUserUtils.getCurrentUserId();
+        if (scope == null || scope.isBlank() || currentUserId == null || currentUserId.isBlank()) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
+        switch (scope.toUpperCase()) {
+            case "CUSTOMER":
+                if (!appointment.getCustomer().getId().equalsIgnoreCase(currentUserId)) {
+                    throw new AppException(ErrorCode.UNAUTHORIZED);
+                }
+                break;
+            case "DOCTOR":
+                if (!appointment.getDoctor().getId().equalsIgnoreCase(currentUserId)) {
+                    throw new AppException(ErrorCode.UNAUTHORIZED);
+                }
+                break;
+            case "MANAGER":
+                break;
+            default:
+                throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
+    }
 
-        Map<LocalDate, Integer> dateToCountMap = tmp.stream()
-                .collect(Collectors.toMap(
-                        AppointmentInNext7DayProjection::getAppointmentDate,
-                        AppointmentInNext7DayProjection::getTotalAppointment
-                ));
+    public Page<AppointmentBasicProjection> searchAppointments(Long stepId, String customerId, String doctorId, LocalDate date, String status, int page, int size) {
+        String scope = CurrentUserUtils.getCurrentScope();
+        if (scope == null || scope.isBlank()) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
+        String currentUserId = CurrentUserUtils.getCurrentUserId();
+        Pageable pageable = PageRequest.of(page, size, Sort.by("appointmentDate").ascending());
+        return switch (scope.toUpperCase()) {
+            case "CUSTOMER" -> appointmentRepository.searchAppointments(stepId, currentUserId, null, date, status, pageable);
+            case "DOCTOR" -> appointmentRepository.searchAppointments(stepId,null, currentUserId, date, status, pageable);
+            case "MANAGER" -> appointmentRepository.searchAppointments(stepId, customerId, doctorId, date, status, pageable);
+            default -> throw new AppException(ErrorCode.ROLE_NOT_EXISTED);
+        };
+    }
 
-        // B2: Duyệt 7 ngày → build list response
-        return IntStream.rangeClosed(0, 6)
-                .mapToObj(i -> {
-                    LocalDate date = startDate.plusDays(i);
-                    return AppointmentInNext7DayResponse.builder()
-                            .appointmentDate(date)
-                            .totalAppointment(dateToCountMap.getOrDefault(date, 0))
-                            .build();
-                })
-                .toList();
+    public AppointmentResponse getAppointmentDetail(Long appointmentId) {
+        String scope = CurrentUserUtils.getCurrentScope();
+        if (scope == null || scope.isBlank()) {
+            throw new AppException(ErrorCode.UNAUTHENTICATED);
+        }
+        String currentUserId = CurrentUserUtils.getCurrentUserId();
+        Appointment appointment = appointmentRepository.findById(appointmentId)
+                .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
+
+        switch (scope) {
+            case "CUSTOMER" -> {
+                if (!appointment.getCustomer().getId().equalsIgnoreCase(currentUserId)) {
+                    throw new AppException(ErrorCode.UNAUTHORIZED);
+                }
+                return appointmentMapper.toAppointmentResponse(appointment);
+            }
+            case "DOCTOR" -> {
+                if (!appointment.getDoctor().getId().equalsIgnoreCase(currentUserId)) {
+                    throw new AppException(ErrorCode.UNAUTHORIZED);
+                }
+                return appointmentMapper.toAppointmentResponse(appointment);
+            }
+            case "MANAGER" -> {
+                return appointmentMapper.toAppointmentResponse(appointment);
+            }
+            default -> throw new AppException(ErrorCode.ROLE_NOT_EXISTED);
+        }
     }
 
     @PreAuthorize("hasRole('DOCTOR') or hasRole('MANAGER')")
@@ -108,22 +146,8 @@ public class AppointmentService {
         return appointmentMapper.toAppointmentResponse(appointmentRepository.save(appointment));
     }
 
-    @PreAuthorize("hasRole('DOCTOR')")
-    public List<AppointmentResponse> getAppointmentWithStatusPendingChangeByDoctorId(String doctorId) {
-        return appointmentRepository.findAllByStatusAndDoctor_Id("PENDING_CHANGE", doctorId)
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
     //auto generate
-    public Appointment createInitialAppointment(
-            User customer,
-            Doctor doctor,
-            LocalDate date,
-            String shift,
-            TreatmentStep treatmentStep
-    ) {
+    public void createInitialAppointment(User customer,Doctor doctor, LocalDate date, String shift, TreatmentStep treatmentStep) {
         Appointment appointment = appointmentRepository.save(Appointment.builder()
                 .customer(customer)
                 .doctor(doctor)
@@ -135,8 +159,6 @@ public class AppointmentService {
                 .createdAt(LocalDate.now())
                 .build());
         reminderService.createReminderForAppointment(appointment);
-        return appointment;
-
     }
 
     public boolean isDoctorAvailable(String doctorId, LocalDate date, String shift) {
@@ -159,24 +181,9 @@ public class AppointmentService {
         return shiftAppointmentCount < 10;
     }
 
-    @PreAuthorize("hasRole('CUSTOMER') or hasRole('MANAGER')")
-    public List<AppointmentResponse> getAppointmentsForCustomer(String customerId) {
-        return appointmentRepository.findAppointmentByCustomerIdAndStatusNot(customerId, "CANCELLED")
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
-    @PreAuthorize("hasRole('DOCTOR') or hasRole('MANAGER')")
-    public List<AppointmentResponse> getAppointmentsForDoctorByDate(String doctorId, LocalDate date) {
-        return appointmentRepository.findAppointmentByDoctorIdAndStatusNotAndAppointmentDate(doctorId, "CANCELLED", date)
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
     //create appointment
     @PreAuthorize("hasRole('DOCTOR') or hasRole('MANAGER')")
+    @Transactional
     public AppointmentResponse createAppointment(AppointmentCreateRequest req) {
         TreatmentStep step = treatmentStepRepository.findById(req.getTreatmentStepId())
                 .orElseThrow(() -> new AppException(ErrorCode.TREATMENT_TYPE_NOT_EXISTED));
@@ -196,6 +203,10 @@ public class AppointmentService {
         if (step.getStatus().equalsIgnoreCase("COMPLETED")
                 || step.getStatus().equalsIgnoreCase("CANCELLED")) {
             throw new AppException(ErrorCode.APPOINTMENT_NOT_CHANGE);
+        }
+        req.setShift(req.getShift().toUpperCase());
+        if (!"MORNING".equalsIgnoreCase(req.getShift()) && !"AFTERNOON".equalsIgnoreCase(req.getShift())) {
+            throw new AppException(ErrorCode.INVALID_SHIFT_VALUE);
         }
 
         Doctor doctor = doctorRepository.findById(req.getDoctorId())
@@ -221,7 +232,11 @@ public class AppointmentService {
     //thay đổi lịch hẹn customer yêu cầu
     @Transactional
     public AppointmentResponse changeAppointmentForCustomer(Long appointmentId, ChangeAppointmentByCustomerRequest request) {
-        Appointment appointment = isAvailableForChange(appointmentId, request.getRequestedDate(), request.getRequestedShift());
+        Appointment appointment = appointmentRepository.findById(appointmentId)
+                .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
+        validateAppointmentAvailableForChange(appointment, request.getRequestedDate(), request.getRequestedShift());
+        validateCanChangeAppointment(appointment);
+
         if (appointment.getTreatmentStep().getStatus().equalsIgnoreCase("COMPLETED")
                 || appointment.getTreatmentStep().getStatus().equalsIgnoreCase("CANCELLED")) {
             throw new AppException(ErrorCode.APPOINTMENT_NOT_CHANGE);
@@ -243,13 +258,16 @@ public class AppointmentService {
     public AppointmentResponse confirmChangeAppointment(Long id, ConfirmChangeAppointmentRequest request) {
         Appointment appointment = appointmentRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
-
+        validateCanChangeAppointment(appointment);
         if (!"PENDING_CHANGE".equalsIgnoreCase(appointment.getStatus())) {
             throw new AppException(ErrorCode.CAN_NOT_BE_UPDATED_STATUS);
         }
-
         switch (request.getStatus().toUpperCase()) {
             case "CONFIRMED":
+                boolean isAvailable = isDoctorAvailable(appointment.getDoctor().getId(), appointment.getRequestedDate(), appointment.getRequestedShift());
+                if (!isAvailable) {
+                    throw new AppException(ErrorCode.DOCTOR_NOT_AVAILABLE);
+                }
                 reminderRepository.deleteByAppointment_Id(appointment.getId());
                 appointment.setAppointmentDate(appointment.getRequestedDate());
                 appointment.setShift(appointment.getRequestedShift());
@@ -267,16 +285,13 @@ public class AppointmentService {
         return appointmentMapper.toAppointmentResponse(appointmentRepository.save(appointment));
     }
 
-    @PreAuthorize("hasRole('MANAGER')")
-    public List<AppointmentResponse> getAllAppointments() {
-        return appointmentRepository.findAllByOrderByAppointmentDateAsc().stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
+    @Transactional
     @PreAuthorize("hasRole('DOCTOR') or hasRole('MANAGER')")
     public AppointmentResponse changeAppointmentForManagerOrDoctor(Long appointmentId, ChangeAppointmentByDoctorOrManagerRequest request) {
-        Appointment appointment = isAvailableForChange(appointmentId, request.getAppointmentDate(), request.getShift());
+        Appointment appointment = appointmentRepository.findById(appointmentId)
+                .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
+        validateCanChangeAppointment(appointment);
+        validateAppointmentAvailableForChange(appointment, request.getAppointmentDate(), request.getShift());
 
         if (appointment.getTreatmentStep().getStatus().equalsIgnoreCase("COMPLETED")
                 || appointment.getTreatmentStep().getStatus().equalsIgnoreCase("CANCELLED")) {
@@ -296,38 +311,14 @@ public class AppointmentService {
         return appointmentMapper.toAppointmentResponse(appointmentRepository.save(appointment));
     }
 
-
-    @PreAuthorize("hasRole('MANAGER')")
-    public List<AppointmentResponse> getAllAppointmentTodayForManager() {
-        return appointmentRepository.findAllByAppointmentDateIsOrderByAppointmentDateAsc(LocalDate.now())
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
     @PreAuthorize("hasRole('DOCTOR') or hasRole('MANAGER')")
     public AppointmentResponse cancelAppointment(AppointmentCancelRequest request) {
         Appointment appointment = appointmentRepository.findById(request.getAppointmentId())
                 .orElseThrow(() -> new AppException(ErrorCode.APPOINTMENT_NOT_FOUND));
+        validateCanChangeAppointment(appointment);
         appointment.setStatus("CANCELLED");
         appointment.setNotes(request.getNote());
         reminderRepository.deleteByAppointment_Id(request.getAppointmentId());
         return appointmentMapper.toAppointmentResponse(appointmentRepository.save(appointment));
-    }
-
-    public List<AppointmentResponse> getAppointmentByStepId(Long stepId) {
-        List<Appointment> appointments = appointmentRepository.findByTreatmentStep_Id(stepId);
-        return appointments
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
-    }
-
-    public List<AppointmentResponse> getAppointmentByDoctorId(String doctorId) {
-        List<Appointment> appointments = appointmentRepository.findAllByDoctor_Id(doctorId);
-        return appointments
-                .stream()
-                .map(appointmentMapper::toAppointmentResponse)
-                .toList();
     }
 }
